requirements.txt
blinker
cachetools
certifi
charset-normalizer
click
contourpy
cycler
filelock
Flask
fonttools
fsspec
grpcio
grpcio-status
gunicorn
httplib2
huggingface-hub
idna
itsdangerous
Jinja2
joblib
kiwisolver
MarkupSafe
matplotlib  
mpmath
networkx
nltk
numpy
opencv-python
packaging
pillow
proto-plus
protobuf
pyasn1
pyasn1-modules
pyparsing
pytesseract
python-dateutil==2.9.0.post0
PyYAML
regex
requests
rsa
safetensors
six
sympy
tokenizers
torch
tqdm
transformers
typing_extensions
uritemplate
urllib3
Werkzeug
flask
flask_cors
numpy
opencv-python
pandas
requests
scipy
seaborn
tensorflow==2.12.0
